--- a/backend/graphing.py
+++ b/backend/graphing.py
@@ -8,9 +8,6 @@ import utils
 # Standard Library Imports
 import json
 
-# Graph Theory Imports
-import networkx as nx
-
 ################################################################################
 # Type Declarations
 ################################################################################