# Changing Graphing Libraries for Performance Reasons #

## 🗣 Description ##
Work was performed to change from NetworkX to RustworkX as our graph theory library due to performance gains.

## 💭 Motivation and context ##

Execution time for building parsers significantly increased as the size of the parser increased. After exploring other libraries, RustworkX was found to dramatically decrease the execution time required for larger parsers.

## 🧪 Testing ##

I ran the parser build execution both with the old NetworkX code and the new RustworkX code and compared the output code. The output code remained the same between both libraries. I also timed the execution of both. For are larger parser, the execution time decreased from 1 minute, 32.059 seconds to 1.412 seconds and for our smaller parser, the execution time decreased from 0.322 seconds to 0.158 seconds.

<!--
## 📷 Screenshots (if appropriate) ##

Uncomment this section if a screenshot is needed.

-->

## ✅ Pre-approval checklist ##

<!-- Remove any of the following that do not apply. -->
<!-- Draft PRs should have one or more unchecked boxes. -->
<!-- If you're unsure about any of these, don't hesitate to ask. -->
<!-- We're here to help! -->

- [ ] This PR has an informative and human-readable title.
- [ ] Changes are limited to a single goal - *eschew scope creep!*
- [ ] *All* future TODOs are captured in issues, which are referenced
      in code comments.
- [ ] All relevant type-of-change labels have been added.
- [ ] I have read the [CONTRIBUTING](../blob/develop/CONTRIBUTING.md) document.
- [ ] These code changes follow [cisagov code standards](https://github.com/cisagov/development-guide).
- [ ] All relevant repo and/or project documentation has been updated
      to reflect the changes in this PR.
- [ ] Tests have been added and/or modified to cover the changes in this PR.
- [ ] All new and existing tests pass.

## ✅ Pre-merge checklist ##

<!-- Remove any of the following that do not apply. -->
<!-- These boxes should remain unchecked until the pull request has been -->
<!-- approved. -->

- [ ] Revert dependencies to default branches.
- [ ] Finalize version.

## ✅ Post-merge checklist ##

<!-- Remove any of the following that do not apply. -->

- [ ] Create a release.